+++ components/ledger/proposal/deposit/Success.tsx
@@ -77,7 +77,7 @@ const useStyles = makeStyles({
 
 const Success = () => {
     const classes = useStyles();
-    const [, setOpen] = React.useState(false);
+    const [open, setOpen] = React.useState(false);